chore: move HashComputation to its own files
